requirements.txt
streamlit
pdfminer.six
spacy
pandas
scikit-learn
nltk
python-docx
openpyxl
